PMM-13042: Moves methods and page objects
