+++ tests/pages/components/queryAnalytics/queryAnalyticsFilters.js
