+++ tests/pages/components/queryAnalytics/queryAnalyticsQueryDetails.js
