+++ tests/pages/queryAnalyticsPage.js
@@ -1,8 +1,8 @@
 const { I } = inject();
-const { QueryAnalyticsFilters } = require('./queryAnalyticsFilters');
-const { QueryAnalyticsData } = require('./queryAnalyticsData');
-const { DashboardLinkContainer } = require('../components/dashboardLinkContainer');
-const { QueryAnalyticsQueryDetails } = require('./queryAnalyticsQueryDetails');
+const { QueryAnalyticsFilters } = require('./components/queryAnalytics/queryAnalyticsFilters');
+const { QueryAnalyticsData } = require('./components/queryAnalytics/queryAnalyticsData');
+const { DashboardLinkContainer } = require('./components/dashboardLinkContainer');
+const { QueryAnalyticsQueryDetails } = require('./components/queryAnalytics/queryAnalyticsQueryDetails');
 
 class QueryAnalyticsPage {
   constructor() {
@@ -13,10 +13,6 @@ class QueryAnalyticsPage {
     this.queryDetails = new QueryAnalyticsQueryDetails();
     this.elements = {
       spinner: locate('//div[@data-testid="Spinner" or @class="preloader"]'),
-      mainMetricsContainer: locate('//div[@data-testid="group-by"]'),
-      selectedMainMetric: () => this.elements.mainMetricsContainer.find('//span[@class="ant-select-selection-item"]'),
-      mainMetricByName: (metricsName) => this.elements.selectedMainMetric().withText(metricsName),
-      mainMetricFromDropdown: (metricName) => locate(`//div[@class="ant-select-item-option-content" and text()="${metricName}"]`),
       metricsSorting: (columnNumber) => locate(`(//a[@data-testid='sort-by-control'])[${columnNumber}]`),
       columnName: (columnName) => locate(`//span[text()="${columnName}"]`),
       clipboardLink: locate(I.getPopUpLocator()).find('span').find('span'),
@@ -34,18 +30,6 @@ class QueryAnalyticsPage {
     I.waitForDetached(this.elements.spinner, 60);
   }
 
-  async changeMainMetric(newMainMetric) {
-    const oldMainMetric = await I.grabTextFrom(this.elements.selectedMainMetric());
-
-    I.click(this.elements.mainMetricsContainer);
-    I.click(this.elements.mainMetricFromDropdown(newMainMetric));
-    I.waitForDetached(this.elements.mainMetricByName(oldMainMetric), 10);
-    I.waitForElement(this.elements.mainMetricByName(newMainMetric), 10);
-  }
-
-  verifyMainMetric(mainMetric) {
-    I.waitForVisible(this.elements.mainMetricByName(mainMetric));
-  }
 
   changeSorting(columnNumber) {
     I.waitForElement(this.elements.metricsSorting(columnNumber), 30);
